# Conflicts:
#	routes/web.php

+++ database/migrations/2025_01_22_132606_add_order_id_column_to_transactions_table.php
@@ -4,7 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class AddOrderIdToTransactionsTable extends Migration
+class AddOrderIdColumnToTransactionsTable extends Migration
 {
     /**
      * Run the migrations.
@@ -13,17 +13,20 @@ class AddOrderIdToTransactionsTable extends Migration
      */
     public function up()
     {
-        Schema::table('transactions', function (Blueprint $table) {
-            // Agregar columna order_id como nullable para evitar problemas al migrar datos existentes
-            $table->unsignedBigInteger('order_id')->nullable()->after('id');
+        // Verificar si la columna 'order_id' ya existe
+        if (!Schema::hasColumn('transactions', 'order_id')) {
+            Schema::table('transactions', function (Blueprint $table) {
+                // Agregar columna order_id como nullable para evitar problemas al migrar datos existentes
+                $table->unsignedBigInteger('order_id')->nullable()->after('id');
 
-            // Crear la relación de clave foránea con orders.id
-            $table->foreign('order_id')
-                ->references('id')
-                ->on('orders')
-                ->onDelete('cascade') // Elimina las transacciones si se elimina el pedido relacionado
-                ->onUpdate('cascade');
-        });
+                // Crear la relación de clave foránea con orders.id
+                $table->foreign('order_id')
+                    ->references('id')
+                    ->on('orders')
+                    ->onDelete('cascade') // Elimina las transacciones si se elimina el pedido relacionado
+                    ->onUpdate('cascade');
+            });
+        }
     }
 
     /**
@@ -33,10 +36,13 @@ class AddOrderIdToTransactionsTable extends Migration
      */
     public function down()
     {
-        Schema::table('transactions', function (Blueprint $table) {
-            // Eliminar la clave foránea y la columna order_id
-            $table->dropForeign(['order_id']);
-            $table->dropColumn('order_id');
-        });
+        // Verificar si la columna 'order_id' existe antes de intentar eliminarla
+        if (Schema::hasColumn('transactions', 'order_id')) {
+            Schema::table('transactions', function (Blueprint $table) {
+                // Eliminar la clave foránea y la columna order_id
+                $table->dropForeign(['order_id']);
+                $table->dropColumn('order_id');
+            });
+        }
     }
 }
